+++ src/main/java/org/bboxdb/network/server/handler/query/HandleBoundingBoxTimeQuery.java
@@ -24,7 +24,7 @@ import org.bboxdb.network.packages.PackageEncodeException;
 import org.bboxdb.network.packages.request.QueryBoundingBoxTimeRequest;
 import org.bboxdb.network.packages.response.ErrorResponse;
 import org.bboxdb.network.server.ClientConnectionHandler;
-import org.bboxdb.network.server.ClientQuery;
+import org.bboxdb.network.server.StreamClientQuery;
 import org.bboxdb.network.server.ErrorMessages;
 import org.bboxdb.storage.entity.TupleStoreName;
 import org.bboxdb.storage.queryprocessor.queryplan.BoundingBoxAndTimeQueryPlan;
@@ -59,7 +59,7 @@ public class HandleBoundingBoxTimeQuery implements QueryHandler {
 			final QueryPlan queryPlan = new BoundingBoxAndTimeQueryPlan(queryRequest.getBoundingBox(), 
 					queryRequest.getTimestamp());
 	
-			final ClientQuery clientQuery = new ClientQuery(queryPlan, queryRequest.isPagingEnabled(), 
+			final StreamClientQuery clientQuery = new StreamClientQuery(queryPlan, queryRequest.isPagingEnabled(), 
 					queryRequest.getTuplesPerPage(), clientConnectionHandler, packageSequence, requestTable);
 			
 			clientConnectionHandler.getActiveQueries().put(packageSequence, clientQuery);
